+++ hpc/nvidia/cuda/fakeCUDA/utils.hpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0
 /* Copyright (c) 2025 Rong Tao */
+#pragma once
 #include <sys/types.h>
-
 
 void memdump(const void *mem, size_t size);
 void debug_memdump(const void *mem, size_t size);